requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
Jinja2==3.1.4
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
fsrs==6.3.1
pytest==8.3.3
httpx==0.27.2
